requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
rich>=13.7.0
click>=8.1.7
python-dateutil>=2.8.2

